lambda/requirements.txt
psycopg2-binary
aws-rds
boto3